+++ Gruntfile.js
@@ -1,4 +1,7 @@
 module.exports = function(grunt) {
+  // Load grunt tasks automatically
+  require('load-grunt-tasks')(grunt);
+
   grunt
       .initConfig({
         pkg : grunt.file.readJSON('package.json'),
@@ -30,9 +33,6 @@ module.exports = function(grunt) {
       });
 
   var stage = grunt.option('stage') || 'dev';
-  
-  grunt.loadNpmTasks('grunt-contrib-watch');
-  grunt.loadNpmTasks('grunt-contrib-compass');
 
   grunt.registerTask('default', [ 'compass:' + stage ]);
 };