+++ capistrano/drupal.rb
@@ -6,8 +6,8 @@ end
 # Backup the database when publishing a new release
 Rake::Task["deploy:publishing"].enhance ["drupal:dbbackup"]
 
-# Copy drush aliases when we check compatibility
-Rake::Task["deploy:check"].enhance ["drush:initialize"]
+# Copy drush aliases after linking the new release
+Rake::Task["deploy:symlink:release"].enhance ["drush:initialize"]
 
 # After publication run updates
 Rake::Task["deploy:published"].enhance do 
@@ -15,33 +15,38 @@ Rake::Task["deploy:published"].enhance do
 end
 
 namespace :drupal do
+  desc "Install Drupal"
+  task :install do
+    invoke 'drush:siteinstall'
+  end
+  
   desc "Copy Drupal and web server configuration files"
   task :settings do
     on roles(:app) do
       fetch(:site_folder).each do |folder|
-        if test " [ -f #{current_path}/public/sites/#{folder}/settings.php ]"
-          execute :rm, "-f", "#{current_path}/public/sites/#{folder}/settings.php"
+        if test " [ -e #{current_path}/#{fetch(:webroot, 'public')}/sites/#{folder}/settings.php ]"
+          execute :rm, "-f", "#{current_path}/#{fetch(:webroot, 'public')}/sites/#{folder}/settings.php"
         end
-        execute :ln, '-s', "#{current_path}/public/sites/#{folder}/settings.#{fetch(:stage)}.php", "#{current_path}/public/sites/#{folder}/settings.php"
+        execute :ln, '-s', "#{current_path}/#{fetch(:webroot, 'public')}/sites/#{folder}/settings.#{fetch(:stage)}.php", "#{current_path}/#{fetch(:webroot, 'public')}/sites/#{folder}/settings.php"
       end
         
       # If a .htaccess file for the stage exists
-      if test " [ -f #{current_path}/public/htaccess.#{fetch(:stage)} ]"
+      if test " [ -f #{current_path}/#{fetch(:webroot, 'public')}/htaccess.#{fetch(:stage)} ]"
         # If there is currently an .htaccess file
-        if test " [ -f #{current_path}/public/.htaccess ]"
-          execute :rm, "#{current_path}/public/.htaccess"
+        if test " [ -f #{current_path}/#{fetch(:webroot, 'public')}/.htaccess ]"
+          execute :rm, "#{current_path}/#{fetch(:webroot, 'public')}/.htaccess"
         end
         
-        execute :ln, '-s', "#{current_path}/public/htaccess.#{fetch(:stage)}", "#{current_path}/public/.htaccess"
+        execute :ln, '-s', "#{current_path}/#{fetch(:webroot, 'public')}/htaccess.#{fetch(:stage)}", "#{current_path}/#{fetch(:webroot, 'public')}/.htaccess"
       end
       
       # If there a robots.txt file for the stage exists
-      if test " [ -f #{current_path}/public/robots.#{fetch(:stage)}.txt ]"
-        if test " [ -f #{current_path}/public/robots.txt ]"
-          execute :rm, "#{current_path}/public/robots.txt"
+      if test " [ -f #{current_path}/#{fetch(:webroot, 'public')}/robots.#{fetch(:stage)}.txt ]"
+        if test " [ -f #{current_path}/#{fetch(:webroot, 'public')}/robots.txt ]"
+          execute :rm, "#{current_path}/#{fetch(:webroot, 'public')}/robots.txt"
         end
       
-        execute :ln, '-s', "#{current_path}/public/robots.#{fetch(:stage)}.txt", "#{current_path}/public/robots.txt"
+        execute :ln, '-s', "#{current_path}/#{fetch(:webroot, 'public')}/robots.#{fetch(:stage)}.txt", "#{current_path}/#{fetch(:webroot, 'public')}/robots.txt"
       end
     end
   end
@@ -53,6 +58,7 @@ namespace :drupal do
       last_release_path = releases_path.join(last_release)
       
       within "#{last_release_path}/public" do
+      within "#{last_release_path}/#{fetch(:app_webroot, 'public')}" do
       	execute :drush, "-y sql-drop -l #{fetch(:site_url)} &&", %{$(drush sql-connect -l #{fetch(:site_url)}) < #{last_release_path}/db.sql}
       end
     end