+++ config/_deploy.rb
@@ -1,14 +1,11 @@
 # The name for the application, should only be things that can be in a directory name
-set :application, 'app'
+set :application, '{{name}}'
 
 # The repo URL
-set :repo_url, 'repo'
-
-# Use git for the SCM
-set :scm, :git
+set :repo_url, '{{repository}}'
 
 # Or use the line below to deploy via rsync
-# set :scm, :rsync
+set :scm, :{{deploy_via}}
 
 # Use a remote cache for git
 set :deploy_via, :remote_cache
@@ -25,11 +22,8 @@ set :user_sudo, false
 # Whether to use a pseudo-TTY
 set :pty, true
 
-# Array of folders to share
-set :linked_dirs, %w{public/sites/default/files}
-
 # Number of release directories to keep
-set :keep_releases, 3
+set :keep_releases, {{keep_releases}}
 
 # Add custom SSH config
 set :ssh_options, {
@@ -37,7 +31,16 @@ set :ssh_options, {
 }
 
 # Platform
-set :platform, "drupal"
+set :platform, "{{platform}}"
+
+{{#is_drupal}}
+set :drupal_features, {{features}}
+set :drupal_cmi, {{cmi}}
+set :drupal_db_updates, true
+
+# Array of folders to share
+set :linked_dirs, %w[{{app_webroot}}/sites/default/files]
+{{/is_drupal}}
 
 # rsync settings
 set :rsync_options, %w[--recursive --chmod=Dug=rwx,Do=rx --perms --delete --delete-excluded --exclude=.git* --exclude=node_modules]